refactor: extract SettingsStore+AIProviderReadiness (#636 slice H)
Move 4 AIProvider readiness computed properties from SettingsStore.swift
(1,625 lines) into a new focused extension:

- hasAPIKey
- hasUsableAIProviderCredential
- aiProviderConfigurationIsReady
- aiProviderCompatibilityIssue

Widen 2 helpers from private to internal (unchanged bodies) so the extension
can reach them:
- credentialIsAvailableForUserAction(value:persistenceState:)
- aiProviderCredentialIsAvailableForCurrentProvider(allowsLegacyOpenAICredential:)

Byte-preserving property bodies. Base file drops to ~1,580 lines.

Closes #901.
